feat(conversation): add OtherUserId to ConversationDTO
Added a new property `OtherUserId` of type `long?` to the `ConversationDTO` class to store the identifier for another user in a conversation. Updated the `GetConversationsQueryHandler` to populate `OtherUserId` based on the conversation type, retrieving the user ID of the other participant for direct conversations or setting it to `null` otherwise.