Add LastMessageType to ConversationDTO and refactor mapping
Added a new `LastMessageType` property to `ConversationDTO` to
store the type of the last message in a conversation. Updated
the `ToDTO` method in `ConversationExtensions` to populate
this property when the `LastMessage` object is not null.

Refactored the `ToDTOs` method to remove redundant logic for
setting `LastMessage` and `LastMessageMetaData`, as this is
now handled in `ToDTO`.

Updated `GetConversationsQueryHandler` to include the
`LastMessageType` property when mapping `Conversation`
objects to DTOs.